optimisation: Do some η-reductions
We take advantage of the previously introduced concept of "arities"
to determine when this optimisation is sound. Arity represents
how many arguments can be passed before the term is no longer in WHNF.

We can't η-reduce when the wrapping lambda has a higher arity, as
that would mean e.g. `\x y z -> f x y z` where `f = \_ -> perror` would be
optimised into `\_ -> perror` which has different behaviour.

We don't catch all cases, specifically, the `f` in `\x -> f x` must be
a hoisted term (therefore closed). It doesn't have to be this restricted,
but then we'd need per-level tracking of variable usage by terms, which is
certainly possible (and also useful for automatic hoisting / CSE) but not
something I'm going to do in this commit.

Fixes #32