flake: fix hercules-ci configuration
Hercules CI was not configured optimally for this repo. I've renamed
nixCi to ciNix and configured it to be consistent with the Hercules CI
template, documentation and our other repos. It is no longer an
attribute that can be evaluated, instead Hercules CI is supposed to
simply build all of our existing flake attribute outputs, instead of
being a set that contains all of them recursively.